Sentences performance test
