+++ python/biomedicus/version.py
@@ -1,0 +1,4 @@
+# coding: utf-8
+# file generated by setuptools_scm
+# don't change, don't track in version control
+version = '0.1.dev4+g34fc875.d20190607'
